+++ src/OpenTelemetry/Logs/Builder/LoggerProviderBuilderExtensions.cs
@@ -26,7 +26,7 @@ namespace OpenTelemetry.Logs;
 /// <summary>
 /// Contains extension methods for the <see cref="LoggerProviderBuilder"/> class.
 /// </summary>
-internal static class LoggerProviderBuilderExtensions
+public static class LoggerProviderBuilderExtensions
 {
     /// <summary>
     /// Sets the <see cref="ResourceBuilder"/> from which the Resource associated with
@@ -151,9 +151,9 @@ internal static class LoggerProviderBuilderExtensions
     /// <returns><see cref="LoggerProvider"/>.</returns>
     public static LoggerProvider Build(this LoggerProviderBuilder loggerProviderBuilder)
     {
-        if (loggerProviderBuilder is LoggerProviderServiceCollectionBuilder loggerProviderServiceCollectionBuilder)
+        if (loggerProviderBuilder is LoggerProviderBuilderBase loggerProviderBuilderBase)
         {
-            return loggerProviderServiceCollectionBuilder.Build();
+            return loggerProviderBuilderBase.Build();
         }
 
         return new LoggerProvider();